feat(06-01): add metadata to all 14 dedicated project pages and dynamic route
