+++ src/app/projects/cac-unit-economics/page.tsx
@@ -1,31 +1,30 @@
-import { headers } from 'next/headers'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { siteConfig } from '@/lib/site'
-import { CACContent } from './_components/CACContent'
+import { Metadata } from 'next'
+import CACPageContent from './_components/CACPageContent'
 
-export default async function CACUnitEconomics() {
-  const nonce = (await headers()).get('x-nonce')
+export const dynamic = 'force-static'
 
-  return (
-    <>
-      <ProjectJsonLd
-        title="CAC & Unit Economics Analysis"
-        description="Comprehensive CAC analysis and LTV:CAC ratio optimization that achieved 32% cost reduction through strategic partner channel optimization."
-        slug="cac-unit-economics"
-        category="Data Analytics"
-        tags={['Customer Acquisition Cost', 'Unit Economics', 'LTV:CAC Ratio', 'Revenue Operations']}
-        nonce={nonce}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: siteConfig.url },
-          { name: 'Projects', url: `${siteConfig.url}/projects` },
-          { name: 'CAC & Unit Economics Analysis', url: `${siteConfig.url}/projects/cac-unit-economics` },
-        ]}
-        nonce={nonce}
-      />
-      <CACContent />
-    </>
-  )
+export const metadata: Metadata = {
+  title: 'Customer Acquisition Cost Optimization & Unit Economics Dashboard | Richard Hudson',
+  description: 'Comprehensive CAC analysis and LTV:CAC ratio optimization that achieved 32% cost reduction through strategic partner channel optimization. Industry-benchmark 3.6:1 efficiency ratio with 8.4-month payback period across multi-tier SaaS products.',
+  openGraph: {
+    title: 'Customer Acquisition Cost Optimization & Unit Economics Dashboard',
+    description: 'Comprehensive CAC analysis and LTV:CAC ratio optimization that achieved 32% cost reduction through strategic partner channel optimization. Industry-benchmark 3.6:1 efficiency ratio with 8.4-month payback period across multi-tier SaaS products.',
+    url: 'https://richardwhudsonjr.com/projects/cac-unit-economics',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Customer Acquisition Cost Optimization & Unit Economics Dashboard' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Customer Acquisition Cost Optimization & Unit Economics Dashboard',
+    description: 'Comprehensive CAC analysis and LTV:CAC ratio optimization that achieved 32% cost reduction through strategic partner channel optimization. Industry-benchmark 3.6:1 efficiency ratio with 8.4-month payback period across multi-tier SaaS products.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/cac-unit-economics',
+  },
+}
+
+export default function CACUnitEconomicsPage() {
+  return <CACPageContent />
 }
