+++ src/app/projects/commission-optimization/page.tsx
@@ -1,31 +1,30 @@
-import { headers } from 'next/headers'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { siteConfig } from '@/lib/site'
-import { CommissionContent } from './_components/CommissionContent'
+import { Metadata } from 'next'
+import CommissionPageContent from './_components/CommissionPageContent'
 
-export default async function CommissionOptimization() {
-  const nonce = (await headers()).get('x-nonce')
+export const dynamic = 'force-static'
 
-  return (
-    <>
-      <ProjectJsonLd
-        title="Commission Optimization"
-        description="Advanced commission management and partner incentive optimization platform managing $254K+ commission structures with 34% performance improvement."
-        slug="commission-optimization"
-        category="Revenue Operations"
-        tags={['Commission Management', 'Sales Compensation', 'Incentive Design', 'Revenue Operations']}
-        nonce={nonce}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: siteConfig.url },
-          { name: 'Projects', url: `${siteConfig.url}/projects` },
-          { name: 'Commission Optimization', url: `${siteConfig.url}/projects/commission-optimization` },
-        ]}
-        nonce={nonce}
-      />
-      <CommissionContent />
-    </>
-  )
+export const metadata: Metadata = {
+  title: 'Commission & Incentive Optimization System | Richard Hudson',
+  description: 'Advanced commission management and partner incentive optimization platform managing $254K+ commission structures. Automated tier adjustments with 23% average commission rate optimization and ROI-driven compensation strategy delivering 34% performance improvement and 87.5% automation efficiency.',
+  openGraph: {
+    title: 'Commission & Incentive Optimization System',
+    description: 'Advanced commission management and partner incentive optimization platform managing $254K+ commission structures. Automated tier adjustments with 23% average commission rate optimization and ROI-driven compensation strategy delivering 34% performance improvement and 87.5% automation efficiency.',
+    url: 'https://richardwhudsonjr.com/projects/commission-optimization',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Commission & Incentive Optimization System' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Commission & Incentive Optimization System',
+    description: 'Advanced commission management and partner incentive optimization platform managing $254K+ commission structures. Automated tier adjustments with 23% average commission rate optimization and ROI-driven compensation strategy delivering 34% performance improvement and 87.5% automation efficiency.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/commission-optimization',
+  },
+}
+
+export default function CommissionOptimizationPage() {
+  return <CommissionPageContent />
 }
