+++ src/app/projects/customer-lifetime-value/page.tsx
@@ -1,31 +1,30 @@
-import { headers } from 'next/headers'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { siteConfig } from '@/lib/site'
-import { CLVContent } from './_components/CLVContent'
+import { Metadata } from 'next'
+import CLVPageContent from './_components/CLVPageContent'
 
-export default async function CustomerLifetimeValueAnalytics() {
-  const nonce = (await headers()).get('x-nonce')
+export const dynamic = 'force-static'
 
-  return (
-    <>
-      <ProjectJsonLd
-        title="Customer Lifetime Value Analysis"
-        description="Advanced CLV analytics platform leveraging BTYD predictive modeling framework with 94.3% prediction accuracy across 5 customer segments."
-        slug="customer-lifetime-value"
-        category="Data Analytics"
-        tags={['Customer Lifetime Value', 'CLV', 'Retention Analytics', 'Revenue Operations']}
-        nonce={nonce}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: siteConfig.url },
-          { name: 'Projects', url: `${siteConfig.url}/projects` },
-          { name: 'Customer Lifetime Value Analysis', url: `${siteConfig.url}/projects/customer-lifetime-value` },
-        ]}
-        nonce={nonce}
-      />
-      <CLVContent />
-    </>
-  )
+export const metadata: Metadata = {
+  title: 'Customer Lifetime Value Predictive Analytics Dashboard | Richard Hudson',
+  description: 'Advanced CLV analytics platform leveraging BTYD (Buy Till You Die) predictive modeling framework. Achieving 94.3% prediction accuracy through machine learning algorithms and real-time customer behavior tracking across 5 distinct customer segments.',
+  openGraph: {
+    title: 'Customer Lifetime Value Predictive Analytics Dashboard',
+    description: 'Advanced CLV analytics platform leveraging BTYD (Buy Till You Die) predictive modeling framework. Achieving 94.3% prediction accuracy through machine learning algorithms and real-time customer behavior tracking across 5 distinct customer segments.',
+    url: 'https://richardwhudsonjr.com/projects/customer-lifetime-value',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Customer Lifetime Value Predictive Analytics Dashboard' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Customer Lifetime Value Predictive Analytics Dashboard',
+    description: 'Advanced CLV analytics platform leveraging BTYD (Buy Till You Die) predictive modeling framework. Achieving 94.3% prediction accuracy through machine learning algorithms and real-time customer behavior tracking across 5 distinct customer segments.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/customer-lifetime-value',
+  },
+}
+
+export default function CustomerLifetimeValuePage() {
+  return <CLVPageContent />
 }
