+++ src/app/projects/partner-performance/page.tsx
@@ -1,31 +1,30 @@
-import { headers } from 'next/headers'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { siteConfig } from '@/lib/site'
-import { PartnerContent } from './_components/PartnerContent'
+import { Metadata } from 'next'
+import PartnerPerformancePageContent from './_components/PartnerPerformancePageContent'
 
-export default async function PartnerPerformanceIntelligence() {
-  const nonce = (await headers()).get('x-nonce')
+export const dynamic = 'force-static'
 
-  return (
-    <>
-      <ProjectJsonLd
-        title="Partner Performance Analytics"
-        description="Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem with real-time performance tracking."
-        slug="partner-performance"
-        category="Partnership Program"
-        tags={['Partner Analytics', 'Channel Performance', 'Partnership Program', 'Revenue Operations']}
-        nonce={nonce}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: siteConfig.url },
-          { name: 'Projects', url: `${siteConfig.url}/projects` },
-          { name: 'Partner Performance Analytics', url: `${siteConfig.url}/projects/partner-performance` },
-        ]}
-        nonce={nonce}
-      />
-      <PartnerContent />
-    </>
-  )
+export const metadata: Metadata = {
+  title: 'Partner Performance Intelligence Dashboard | Richard Hudson',
+  description: 'Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem. Real-time performance tracking following industry-standard 80/20 partner revenue distribution.',
+  openGraph: {
+    title: 'Partner Performance Intelligence Dashboard',
+    description: 'Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem. Real-time performance tracking following industry-standard 80/20 partner revenue distribution.',
+    url: 'https://richardwhudsonjr.com/projects/partner-performance',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Partner Performance Intelligence Dashboard' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Partner Performance Intelligence Dashboard',
+    description: 'Strategic channel analytics and partner ROI intelligence demonstrating 83.2% win rate across multi-tier partner ecosystem. Real-time performance tracking following industry-standard 80/20 partner revenue distribution.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/partner-performance',
+  },
+}
+
+export default function PartnerPerformancePage() {
+  return <PartnerPerformancePageContent />
 }
