+++ src/app/projects/partnership-program-implementation/page.tsx
@@ -1,114 +1,30 @@
-'use client'
+import { Metadata } from 'next'
+import PartnershipProgramPageContent from './_components/PartnershipProgramPageContent'
 
-import { Zap, Clock, Settings, Users } from 'lucide-react'
+export const dynamic = 'force-static'
 
-import { ProjectPageLayout } from '@/components/projects/project-page-layout'
-import { MetricsGrid } from '@/components/projects/metrics-grid'
-import { SectionCard } from '@/components/ui/section-card'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { AchievementsGrid } from './_components/AchievementsGrid'
-import { TechnicalDetails } from './_components/TechnicalDetails'
-import { NarrativeSections } from './_components/NarrativeSections'
+export const metadata: Metadata = {
+  title: 'Enterprise Partnership Program Implementation | Richard Hudson',
+  description: 'Led comprehensive design and implementation of a company\'s first partnership program, creating automated partner onboarding, commission tracking, and performance analytics. Built production-ready integrations with CRM, billing systems, and partner portals.',
+  openGraph: {
+    title: 'Enterprise Partnership Program Implementation',
+    description: 'Led comprehensive design and implementation of a company\'s first partnership program, creating automated partner onboarding, commission tracking, and performance analytics. Built production-ready integrations with CRM, billing systems, and partner portals.',
+    url: 'https://richardwhudsonjr.com/projects/partnership-program-implementation',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Enterprise Partnership Program Implementation' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Enterprise Partnership Program Implementation',
+    description: 'Led comprehensive design and implementation of a company\'s first partnership program, creating automated partner onboarding, commission tracking, and performance analytics. Built production-ready integrations with CRM, billing systems, and partner portals.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/partnership-program-implementation',
+  },
+}
 
-export default function PartnershipProgramPage() {
-  // Standardized metrics configuration
-  const metrics = [
-    {
-      id: 'automation-rate',
-      icon: Zap,
-      label: 'Automation',
-      value: '90%+',
-      subtitle: 'Process Coverage',
-      variant: 'success' as const,
-    },
-    {
-      id: 'launch-time',
-      icon: Clock,
-      label: 'Launch Time',
-      value: '6 mo',
-      subtitle: 'Concept to Production',
-      variant: 'primary' as const,
-    },
-    {
-      id: 'integrations',
-      icon: Settings,
-      label: 'Integrations',
-      value: 'Full Stack',
-      subtitle: 'CRM, Billing, Portals',
-      variant: 'secondary' as const,
-    },
-    {
-      id: 'onboarding',
-      icon: Users,
-      label: 'Onboarding',
-      value: 'Automated',
-      subtitle: 'Partner Journey',
-      variant: 'primary' as const,
-    },
-  ]
-
-  return (
-    <>
-      <ProjectJsonLd
-        title="Enterprise Partnership Program Implementation"
-        description="Led comprehensive design and implementation of a company's first partnership program, creating automated partner onboarding, commission tracking, and performance analytics."
-        slug="partnership-program-implementation"
-        category="Revenue Operations"
-        tags={[
-          'Partnership Program',
-          'Channel Operations',
-          'Partner Onboarding',
-          'Commission Automation',
-        ]}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: 'https://richardwhudsonjr.com' },
-          { name: 'Projects', url: 'https://richardwhudsonjr.com/projects' },
-          { name: 'Enterprise Partnership Program Implementation', url: 'https://richardwhudsonjr.com/projects/partnership-program-implementation' },
-        ]}
-      />
-
-      <ProjectPageLayout
-        title="Enterprise Partnership Program Implementation"
-        description="Led comprehensive design and implementation of a company's first partnership program, creating automated partner onboarding, commission tracking, and performance analytics. Built production-ready integrations with CRM, billing systems, and partner portals."
-        tags={[
-          { label: 'First Partner Program', variant: 'primary' },
-          { label: '90%+ Automation', variant: 'secondary' },
-          { label: 'Full Production', variant: 'primary' },
-          { label: 'Strategic Success', variant: 'secondary' },
-        ]}
-      >
-        {/* Key Metrics using standardized MetricsGrid */}
-        <MetricsGrid metrics={metrics} columns={4} className="mb-8" />
-
-        {/* Key Achievements */}
-        <SectionCard
-          title="Key Achievements"
-          description="Strategic implementation results and business impact"
-          className="mb-8"
-        >
-          <AchievementsGrid />
-        </SectionCard>
-
-        {/* Technical Implementation */}
-        <SectionCard
-          title="Technical Implementation"
-          description="Architecture and integration details"
-          className="mb-8"
-        >
-          <TechnicalDetails />
-        </SectionCard>
-
-        {/* Professional Narrative Sections - STAR Method */}
-        <SectionCard
-          title="Project Narrative"
-          description="Comprehensive case study following the STAR methodology"
-        >
-          <NarrativeSections />
-        </SectionCard>
-      </ProjectPageLayout>
-    </>
-  )
+export default function PartnershipProgramImplementationPage() {
+  return <PartnershipProgramPageContent />
 }
