+++ src/app/projects/revenue-operations-center/page.tsx
@@ -1,31 +1,30 @@
-import { headers } from 'next/headers'
-import { ProjectJsonLd } from '@/components/seo/json-ld/project-json-ld'
-import { BreadcrumbListJsonLd } from '@/components/seo/json-ld/breadcrumb-json-ld'
-import { siteConfig } from '@/lib/site'
-import { RevOpsCenterContent } from './_components/RevOpsCenterContent'
+import { Metadata } from 'next'
+import RevenueOpsCenterPageContent from './_components/RevenueOpsCenterPageContent'
 
-export default async function RevenueOperationsCenter() {
-  const nonce = (await headers()).get('x-nonce')
+export const dynamic = 'force-static'
 
-  return (
-    <>
-      <ProjectJsonLd
-        title="Revenue Operations Center"
-        description="Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs with 96.8% forecast accuracy."
-        slug="revenue-operations-center"
-        category="Revenue Operations"
-        tags={['Revenue Operations', 'RevOps Dashboard', 'Sales Analytics', 'Business Intelligence']}
-        nonce={nonce}
-      />
-      <BreadcrumbListJsonLd
-        items={[
-          { name: 'Home', url: siteConfig.url },
-          { name: 'Projects', url: `${siteConfig.url}/projects` },
-          { name: 'Revenue Operations Center', url: `${siteConfig.url}/projects/revenue-operations-center` },
-        ]}
-        nonce={nonce}
-      />
-      <RevOpsCenterContent />
-    </>
-  )
+export const metadata: Metadata = {
+  title: 'Revenue Operations Command Center | Richard Hudson',
+  description: 'Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs. Real-time insights with 96.8% forecast accuracy and 89.7% operational efficiency across sales, marketing, and partner channels.',
+  openGraph: {
+    title: 'Revenue Operations Command Center',
+    description: 'Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs. Real-time insights with 96.8% forecast accuracy and 89.7% operational efficiency across sales, marketing, and partner channels.',
+    url: 'https://richardwhudsonjr.com/projects/revenue-operations-center',
+    siteName: 'Richard Hudson',
+    images: [{ url: 'https://richardwhudsonjr.com/og-image.png', width: 1200, height: 630, alt: 'Revenue Operations Command Center' }],
+    type: 'article',
+  },
+  twitter: {
+    card: 'summary_large_image',
+    title: 'Revenue Operations Command Center',
+    description: 'Comprehensive revenue operations dashboard consolidating pipeline health, forecasting accuracy, partner performance, and operational KPIs. Real-time insights with 96.8% forecast accuracy and 89.7% operational efficiency across sales, marketing, and partner channels.',
+    images: ['https://richardwhudsonjr.com/og-image.png'],
+  },
+  alternates: {
+    canonical: 'https://richardwhudsonjr.com/projects/revenue-operations-center',
+  },
+}
+
+export default function RevenueOperationsCenterPage() {
+  return <RevenueOpsCenterPageContent />
 }
